Move serde test back to `tests.rs`
